requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.24.0
pandas==2.1.0
requests==2.32.0
python-dotenv==1.1.1
